+++ R/vector_logic_linter.R
@@ -99,6 +99,7 @@ vector_logic_linter <- function() {
       and not(preceding-sibling::expr[last()]/SYMBOL_FUNCTION_CALL[not(text() = 'subset' or text() = 'filter')])
       and not(preceding-sibling::OP-LEFT-BRACKET)
       and not(preceding-sibling::*[not(self::COMMENT)][2][self::SYMBOL_SUB and text() = 'circular'])
+      and not(ancestor::expr[FUNCTION or OP-LAMBDA])
     ]
     /*[not(self::COMMENT)][2]
   "